+++ CHANGELOG.md
@@ -30,6 +30,9 @@
 - [10753](https://github.com/vegaprotocol/vega/issues/10753) - Handle the case that a submitted order is `FoK` in isolated margin to not double discount it from position.
 - [10136](https://github.com/vegaprotocol/vega/issues/10136) - Assure opening auction uncrossing price gets registered in the perps engine.
 - [10727](https://github.com/vegaprotocol/vega/issues/10727) - Allow for a 0 funding rate scaling factor.
+- [10785](https://github.com/vegaprotocol/vega/issues/10785) - Oracles data is getting increasingly slow.
+- [10780](https://github.com/vegaprotocol/vega/issues/10780) - REST API does not handle ID arrays correctly.
+- [10792](https://github.com/vegaprotocol/vega/issues/10792) - Use correct market for market proposer bonus for spots. 
 
 ## 0.74.3